+++ NEWS.md
@@ -10,6 +10,7 @@
 
 * `predictive_accuracy()`, which calculates the predictive accuracy of linear or logistic regression models.
 * `log_loss()` to compute the log-loss of models with binary outcome. The log-loss is a proper scoring function comparable to the `rmse()`.
+* `correct_predictions()` to calculate the expected percentage of correct predictions for models with binary outcome.
 
 ## Bug fixes